refactor(ertp-ledgerguise): improve code organization
- Promote SettlementFacet to src/settlement.ts (was inline in test)
- Extract gnucash-tools.ts with table types and snapshot helpers
- Deprecate escrow.ts more strongly (@deprecated JSDoc)
- Update README with documentation section
- Update CONTRIBUTING.md status (mark completed items)
