+++ packages/ertp-ledgerguise/src/index.ts
@@ -49,6 +49,9 @@ export type {
 export { asGuid } from './guids.js';
 export { makeChartFacet } from './chart.js';
 export { makeErtpEscrow } from './escrow-ertp.js';
+export { makeSettlementFacet } from './settlement.js';
+export type { SettlementFacet, SettlementResult } from './settlement.js';
+/** @deprecated Use {@link makeErtpEscrow} instead */
 export { makeEscrow } from './escrow.js';
 export { wrapBetterSqlite3Database } from './sqlite-shim.js';
 export type { SqlDatabase, SqlStatement } from './sql-db.js';
